improved log output when steam init fails due to extraQL being started from the steam library

+++ source/Servlets.cs
@@ -51,7 +51,7 @@ namespace ExtraQL
 
     public string QuakeSteamFolder { get; set; }
 
-    public int SteamAppId { get { return steamworks.AppID; } set { steamworks.AppID = value; } }
+    public ulong SteamAppId { get { return steamworks.AppID; } set { steamworks.AppID = value; } }
 
     public ThreadStart BringToFrontHandler { get; set; }
